requirements.txt
Flask>=3.1.2
pdf2docx
pdfminer.six
python-docx
pytesseract
pdf2image
Pillow
gunicorn
